// ProjectPos.FrontOffice.FrontEnd/src/proxy/enums/role.ts
export enum Role {
    Admin = "Admin",
    Supervisor = "Supervisor",
    Manager = "Manager",
    SysAdmin = "SysAdmin",
    Cashier = "Cashier",
    Accountant = "Accountant",
    AccountsClerk = "AccountsClerk",
    DebtorsController = "DebtorsController",
    FinanceManager = "FinanceManager",
    SysDeveloper = "SysDeveloper",
    StockController = "StockController",
    ReceivingClerk = "ReceivingClerk",
}